Student Result Marks and Grade
In this java program, we first ask user to enter number of subjects and store it in variable "count". Then using a for loop, we take marks of "count" subjects as input from user and add them to variable "totalMarks". and finally the total mark to compare by using if and else if conditions for 'A' & 'B' Grade...